Update top-level and global gitignore files
The top-level file adds some directories in $HOME to the default
ignores, since those should never be considered for commit to dotfiles.
This mostly is of use, when using tools like ripgrep to quickly ignore
such files.

This also updates the .gitignore.global to current versions of the
upstream github.com/github/gitignore repository.