[NeoVim] Support file detection for Ansible LSP
The Ansible LSP server by default attaches only to files with filetype
'yaml.ansible'. NeoVim doesn't detect this file type out of the box, so
we provide a few file patterns of typical Ansible project directories to
assign the correct type. In addition the YAML LSP has to be made aware
of this filetype too, since the Ansible LSP is designed to work in
conjunction with the YAML LSP.