[NeoVim] Disable treesitter for YAML file indentation
Something is currently broken with the way treesitter tries to indent
YAML files while editing. It always ends up indenting new lines in the
wrong way, leading to a non-pleasant editing experience. Until we find
out how to fix this (in case our configuration is the problem), we
simply disable indentation via treesitter.